chore(deps): update sonyflake to v2.2.0

+++ settings/sonyflake.go
@@ -4,7 +4,7 @@ import "time"
 
 type Sonyflake struct {
 	StartTime time.Time
-	MachineID uint16
+	MachineID int
 }
 
 var SonyflakeSettings = &Sonyflake{}
